Moved classes to dto package to reduce SAT warning
