+++ bundles/org.openhab.voice.googletts/src/main/java/org/openhab/voice/googletts/internal/dto/SynthesizeSpeechRequest.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.voice.googletts.internal.protocol;
+package org.openhab.voice.googletts.internal.dto;
 
 /**
  * Synthesizes speech synchronously: receive results after all text input has been processed.
